third_party/WebKit/LayoutTests/fast/html/imports/import-script-prototypes-expected.txt

This test ensures each import and master is in same JS context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stringInMaster.__proto__ is stringInImport.__proto__
PASS thisDocument !== document is true
PASS thisDocument.__proto__ is document.__proto__
PASS thisDocument.head.__proto__ is document.head.__proto__
PASS (new Image()).ownerDocument is document
PASS thisDocument.createElement('div').ownerDocument is thisDocument
PASS document.createElement('div').ownerDocument is document
PASS successfullyParsed is true

TEST COMPLETE
